+++ BitmovinAnalyticsCollector/Classes/AVPlayer/Classes/AVPlayerAdapter.swift
@@ -207,7 +207,6 @@ class AVPlayerAdapter: NSObject, PlayerAdapter {
     }
     
     func onPlayAttemptFailed(withReason reason: String = VideoStartFailedReason.unknown) {
-        print("onPlayAttemptFailed with Reason \(reason)")
         videoStartFailed = true
         videoStartFailedReason = reason
         stateMachine.transitionState(destinationState: .playAttemptFailed, time: self.player.currentTime())