+++ BitmovinAnalyticsCollector/Classes/Collector/adapters/PlayerAdapter.swift
@@ -4,5 +4,7 @@ import Foundation
 protocol PlayerAdapter {
     func createEventData() -> EventData
     func startMonitoring()
+    func stopMonitoring()
+    func destroy()
     var currentTime: CMTime? { get }
 }